Fix problem with USB debug interface
The problem was that one needs to STATICALLY allocate the buffer
that sends the data to the GPS. In this case, making the rxbuf
static inside gps/main.c fixed the problem.

For safety, the buffer inside usbdbg_puts() was also made static, and aligned
to a word (32-bit) boundary, as advised by the SiLabs USB device library.