gps: Found and fixed the issue about needing a double ON_OFF pulse
This was actually wrong, the Antenova M10478 GPS module doesn't need a double
ON_OFF pin pulse, it just needs an ~150ms delay from the time its EXT_RST pin
is asserted to the time it can dump out data.

What was happening before is that we didn't have a delay in the gps_init
function before sending the on_off pulse, and this meant that the pulse was
missed by the module, which never turned on.

The added delay fixes this issue and we now need only a single pulse.