More trials with GPS optimization
A bug was fixed in gps_parse_nmea(), whereby the wrong buffer was
given as a param to strlen(), which is used in gps_parse_nmea() to
provide a buffer size parameter to nmea_parse.

After this, the processor now enters a hard fault ,probably due to
segmentation fault. More debug is needed probably with a different
board to find out why.